requirements.txt
Flask
request
random
os
gunicron